+++ CHANGELOG.rst
@@ -1,5 +1,5 @@
 ..
-	Copyright 2019-2023 MicroEJ Corp. All rights reserved.
+	Copyright 2019-2024 MicroEJ Corp. All rights reserved.
 	Use of this source code is governed by a BSD-style license that can be found with this software.
 
 =========
@@ -9,6 +9,22 @@ Changelog
 All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_, and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
+
+-------------------
+2.11.0 - 2024-02-22
+-------------------
+
+Changed
+=======
+
+- MicroEJ Multi-Sandbox C Validation (C) 1.1.0 (see `CHANGELOG <tests/llkernel/CHANGELOG.rst>`_).
+- MicroEJ UI Validation (Java) 1.7.0 (see `CHANGELOG <tests/ui/ui3/java-testsuite-runner-ui3/CHANGELOG.rst>`_).
+- MicroEJ VG Validation (Java) 2.6.0 (see `CHANGELOG <tests/vg/java-testsuite-runner-vg/CHANGELOG.rst>`_).
+
+Fixed
+=====
+
+- Fix an include path in CORE C sources.
 
 -------------------
 2.10.0 - 2023-12-05
@@ -27,7 +43,7 @@ Changed
 =======
 
 - MicroEJ UI Validation (Java) 1.4.0 (see `CHANGELOG <tests/ui/ui3/java-testsuite-runner-ui3/CHANGELOG.rst>`_).
-- MicroEJ VG Validation (Java) 2.5.0 (see `CHANGELOG <tests/ui/ui3/java-testsuite-runner-vg/CHANGELOG.rst>`_).
+- MicroEJ VG Validation (Java) 2.5.0 (see `CHANGELOG <tests/vg/java-testsuite-runner-vg/CHANGELOG.rst>`_).
 - MicroEJ SSL Validation (Java) 1.0.2 (see `CHANGELOG <tests/ssl/java-testsuite-runner-ssl/CHANGELOG.rst>`_).
 - MicroEJ NET Validation (Java) 1.0.2 (see `CHANGELOG <tests/net/java-testsuite-runner-net/CHANGELOG.rst>`_).
 - Platform Configuration Additions 2.1.0 (see `CHANGELOG <framework/platform/content-sdk-5/build/CHANGELOG.md>`_).