+++ tests/llkernel/CHANGELOG.rst
@@ -6,6 +6,14 @@ Changelog <https://keepachangelog.com/en/1.0.0/>`__, and this project
 adheres to `Semantic
 Versioning <https://semver.org/spec/v2.0.0.html>`__.
 
+[1.1.0] - 2023-12-13
+--------------------
+
+Added
+~~~~~
+
+- Added new test on ``LLKERNEL_IMPL_copyToROM()`` function.
+
 [1.0.0] - 2023-10-12
 --------------------