+++ tests/vg/java-testsuite-runner-vg/CHANGELOG.rst
@@ -11,6 +11,14 @@ Relationship
 
 See `VG Readme <../README.rst>`_ to have more information about the relationship between the version of the test suite and the VG Pack.
 
+[1.2.0] - 2024-02-15
+--------------------
+
+Changed
+~~~~~~~
+
+- VG Test Suite 2.6.0 (compatibility with VG Pack 1.5.0).
+
 [1.1.0] - 2023-12-05
 --------------------
 
@@ -33,5 +41,5 @@ Added
 - Add the Java Test Suite runner in this PQT repository.
 
 ..
-    Copyright 2023 MicroEJ Corp. All rights reserved.
+    Copyright 2023-2024 MicroEJ Corp. All rights reserved.
     Use of this source code is governed by a BSD-style license that can be found with this software.